// src/typings/utils.d.ts
export interface TableColumnType {
  /** Constraint type (e.g., PRIMARY KEY) */
  readonly constraint: string | null;
  /** Column name */
  readonly name: string | null;
  /** Default value for the column */
  readonly defaultValue: string | null;
  /** Whether the column can be null */
  readonly nullable: boolean;
  /** Data type of the column */
  readonly type: string;
  /** Special attributes (e.g., enum values) */
  readonly special: string | null;
  /** Column comment */
  readonly comment: string | null;
}

/**
 * Interface representing element types of a table column.
 */
export interface TableElementType {
  /** Column name */
  readonly columnName: string;
  /** Data type in lowercase */
  readonly dataType: string;
  /** User-defined type name */
  readonly udtName: string;
  /** Element type for array columns */
  readonly elementType: string | null;
  /** Whether the column is an enum type */
  readonly isEnum: boolean;
  /** Enum values if applicable */
  readonly enumData: string[] | null;
  /** Whether the column is a domain type */
  readonly isDomain: boolean;
  /** Domain information if applicable */
  readonly domainData: DomainTypeData | null;
  /** Whether the column is a composite type */
  readonly isComposite: boolean;
  /** Composite information if applicable */
  readonly compositeData: CompositeTypeData | null;
}
// ...
export interface ExclusiveColumnInfo {
  /** Name of the table. */
  name: string;
  /** Detailed information about the column. */
  column: TableColumnType;
  /** Element type information for the column. */
  element: TableElementType;
  /** Additional information about the column, such as constraints or comments. */
  info: TableColumnRaw;
}

/**
 * Interface representing composite type data for a table column.
 */
export interface CompositeTypeData {
  /** Name of the composite type */
  typeName: string;
  /** Array of attribute names in the composite type */
  attributeNames: string;
  /** Array of attribute types in the composite type */
  attributeTypes: string;
}

/**
 * Interface representing domain type data for a table column.
 */
export interface DomainTypeData {
  /** Name of the domain type */
  domainName: string;
  /** Base type of the domain */
  baseType: string;
  /** Array of constraints applied to the domain */
  constraints: Array<{
    /** Name of the constraint */
    name: string;
    /** Check expression for the constraint */
    checkExpression?: string;
    /** Whether the constraint enforces not null */
    notNull?: boolean;
    /** Default value for the constraint */
    default?: string;
  }>;
}